requirements.txt
djangorestframework==3.14.0
Django==4.1.3
django-rest-knox==4.2.0
